+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 ### Changed
 - Improve graceful shutdown of Google IAP receipt processor.
 - If In-App Purchases validation contain mismatching userIDs, do not return an error.
+- Better handling of matchmaker operations while the interval process is running.
 
 ### Fixed
 - Consistent validation of override operator in runtime leaderboard record writes.